fix: Feature flag without variants cause FunctionClause Error
We were experimenting using Unleash feature variants, and we
noticed when there's not variants the Unleash client raises
an FunctionClause exception.

It can happen when:

The code that is asking for a variant is live

Then:

You're creating a new feature flag, there's a small period the
flag exists with no variants. So the exception happens.

Or you delete all variants of a feature flag for whatever
reason. The exception happens.

Expected behaviour:

It should not raise exception and instead work as the variant is
disabled.

Workarounds: You make sure the feature flag is disabled when
creating or manipulate. But it's error prone relying on the user.

Bonus: The test_helper change is for work on newer versions of Elixir,
where already started apps returns an error.